@@ -40,15 +40,15 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^9.1.11",
-    "@angular/common": "^9.1.11",
-    "@angular/compiler": "^9.1.11",
-    "@angular/core": "^9.1.11",
-    "@angular/elements": "^9.1.11",
-    "@angular/forms": "^9.1.11",
-    "@angular/platform-browser": "^9.1.11",
-    "@angular/platform-browser-dynamic": "^9.1.11",
-    "@angular/router": "^9.1.11",
+    "@angular/animations": "^10.0.8",
+    "@angular/common": "^10.0.8",
+    "@angular/compiler": "^10.0.8",
+    "@angular/core": "^10.0.8",
+    "@angular/elements": "^10.0.8",
+    "@angular/forms": "^10.0.8",
+    "@angular/platform-browser": "^10.0.8",
+    "@angular/platform-browser-dynamic": "^10.0.8",
+    "@angular/router": "^10.0.8",
     "@ngx-loading-bar/core": "^4.2.0",
     "@ngx-loading-bar/http-client": "^4.2.0",
     "@stackblitz/sdk": "^1.2.0",
@@ -65,19 +65,19 @@
     "process": "^0.11.10",
     "quick-lru": "^4.0.1",
     "remark-inline-links": "^3.1.3",
-    "rxjs": "^6.5.5",
+    "rxjs": "^6.6.2",
     "smoothscroll-polyfill": "^0.4.4",
-    "tslib": "^1.10.0",
+    "tslib": "^2.0.0",
     "url-resolve": "^1.0.0",
     "vfile-reporter": "^6.0.0",
     "zone.js": "~0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.901.8",
-    "@angular-devkit/build-ng-packagr": "~0.901.8",
-    "@angular/cli": "^9.1.8",
-    "@angular/compiler-cli": "^9.1.11",
-    "@angular/language-service": "^9.1.11",
+    "@angular-devkit/build-angular": "~0.1000.5",
+    "@angular-devkit/build-ng-packagr": "~0.1000.5",
+    "@angular/cli": "^10.0.5",
+    "@angular/compiler-cli": "^10.0.8",
+    "@angular/language-service": "^10.0.8",
     "@swimlane/prettier-config-swimlane": "^3.0.2",
     "@types/jasmine": "^3.5.10",
     "@types/jasminewd2": "~2.0.3",
@@ -85,12 +85,12 @@
     "@types/node": "^13.1.4",
     "angular-cli-ghpages": "^0.6.2",
     "chg": "^0.4.0",
-    "codelyzer": "^5.2.1",
+    "codelyzer": "^6.0.0",
     "cpx": "^1.5.0",
     "cross-env": "^7.0.2",
     "cypress": "^4.3.0",
     "husky": "^4.2.4",
-    "ng-packagr": "^9.0.2",
+    "ng-packagr": "^10.0.0",
     "node-sass": "^4.13.0",
     "npm-run-all": "^4.1.5",
     "prettier": "^2.0.4",
@@ -98,11 +98,11 @@
     "scss-bundle": "^3.1.1",
     "start-server-and-test": "^1.10.11",
     "ts-node": "^8.8.2",
-    "tslint": "^6.1.1",
+    "tslint": "~6.1.0",
     "tslint-config-prettier": "^1.18.0",
     "tslint-config-security": "^1.16.0",
     "tslint-config-swimlane": "^5.0.0",
-    "typescript": "^3.8.3"
+    "typescript": "^3.9.7"
   },
   "repository": {
     "type": "git",
